split composite metadata into multiple files
